# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Up2You API Documentation",
        "description": "Documentazione delle API per la gestione di eventi e iscritti",
        "version": "1.0.0"
    },
    "paths": {
        "/api/attendees": {
            "get": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Recupera la lista di tutti gli iscritti",
                "operationId": "07252fa9523cadea6324ea21430db788",
                "responses": {
                    "200": {
                        "description": "Lista degli iscritti",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "firstname": {
                                                "type": "string",
                                                "example": "Mario"
                                            },
                                            "lastname": {
                                                "type": "string",
                                                "example": "Rossi"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "mario.rossi@example.com"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Crea un nuovo iscritto",
                "operationId": "1b7f6aec6b49c9767416d8bddd81519b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Mario"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Rossi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mario.rossi@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Iscritto creato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "Mario"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Rossi"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "mario.rossi@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errore di validazione"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/attendees/{id}": {
            "get": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Recupera i dettagli di un iscritto",
                "operationId": "b1ede3ccf9c8db6b7084bea2c3b0b34b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dell'iscritto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dettagli dell'iscritto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "Mario"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Rossi"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "mario.rossi@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Iscritto non trovato"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Aggiorna un iscritto esistente",
                "operationId": "42b6aa1e8b15a60617697ffd254a656e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dell'iscritto da aggiornare",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Mario"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Rossi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mario.rossi@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Iscritto aggiornato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "Mario"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Rossi"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "mario.rossi@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errore di validazione"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Elimina un iscritto",
                "operationId": "737979a7a5f77aaaccf1e47a60af5469",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dell'iscritto da eliminare",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Iscritto eliminato con successo"
                    },
                    "404": {
                        "description": "Iscritto non trovato"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "API Key necessaria per accedere alle API",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}